5758 - Remove validateDescriptionText to use queue instead
